refactor(shared): move persona/rewards runtime values to constants/utils
Relocate the runtime values and helper functions that lived in interface
files into constants/ and utils/, per the LFXV2-2570 placement
convention (interfaces keep types only):

- persona.interface.ts: VALID_PERSONAS, BOARD_SCOPED_PERSONAS,
  PROJECT_SCOPED_PERSONAS -> constants/persona.constants.ts;
  isBoardScopedPersona, isProjectScopedPersona -> utils/persona.utils.ts (new)
- rewards.interface.ts: REWARD_CATEGORIES -> constants/rewards.constants.ts
  (RewardPromotionCategory stays and derives from it type-only)

persona.constants.ts previously imported the persona value-sets from
interfaces; that cross-layer import is now removed since the values live
alongside it. Consumer imports are split so values come from
/constants, functions from /utils, and types from /interfaces. No
behavior change.

Refs: LFXV2-2573
