+++ packages/shared/src/interfaces/persona.interface.ts
@@ -5,20 +5,6 @@ import type { Account } from './account.interface';
 
 export type PersonaType = 'contributor' | 'maintainer' | 'board-member' | 'executive-director';
 
-export const VALID_PERSONAS: ReadonlySet<string> = new Set<string>(['contributor', 'maintainer', 'board-member', 'executive-director']);
-
-export const BOARD_SCOPED_PERSONAS: ReadonlySet<PersonaType> = new Set(['board-member', 'executive-director']);
-
-export function isBoardScopedPersona(persona: PersonaType): boolean {
-  return BOARD_SCOPED_PERSONAS.has(persona);
-}
-
-export const PROJECT_SCOPED_PERSONAS: ReadonlySet<PersonaType> = new Set(['maintainer', 'contributor']);
-
-export function isProjectScopedPersona(persona: PersonaType): boolean {
-  return PROJECT_SCOPED_PERSONAS.has(persona);
-}
-
 export interface PersistedPersonaState {
   primary: PersonaType;
   all: PersonaType[];
